Deduplicate within-line reduction passes
