+++ PyRoute/DeltaPasses/AuxiliaryLineReduce.py
@@ -4,20 +4,13 @@ Created on Jul 13, 2023
 @author: CyberiaResurrection
 """
 from PyRoute.DeltaDebug.DeltaReduce import DeltaReduce
+from PyRoute.DeltaPasses.WithinLineReducer import WithinLineReducer
 from PyRoute.DeltaStar import DeltaStar
 
 
-class AuxiliaryLineReduce(object):
+class AuxiliaryLineReduce(WithinLineReducer):
 
     reducer: DeltaReduce
-
-    def __init__(self, reducer):
-        self.reducer = reducer
-
-    def preflight(self):
-        if self.reducer is not None and self.reducer.sectors is not None and 0 < len(self.reducer.sectors.lines):
-            return True
-        return False
 
     def run(self):
         # build substitution list - reduce _everything_
@@ -80,13 +73,3 @@ class AuxiliaryLineReduce(object):
                 segment.extend(chunk)
 
         self.reducer.sectors = best_sectors
-
-    def _assemble_segment(self, unreduced_lines, subs_list):
-        nu_list = []
-
-        for line in subs_list:
-            if line[0] in unreduced_lines:
-                continue
-            nu_list.append((line[0], line[1]))
-
-        return nu_list
